+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - 1.0.0-beta1
+    - Add FASTQ dataset extractor
     - FASTA extractor now handles comments